refactor(): 클래스 및 메서드 명 변경
