+++ src/main/java/blackjack/domain/role/DealerDrawChoice.java
@@ -2,7 +2,7 @@ package blackjack.domain.role;
 
 import java.util.Random;
 
-public class DealerDrawable {
+public class DealerDrawChoice {
 
 	private static final Random random = new Random();
