+++ src/test/java/blackjack/factory/CardMockFactory.java
@@ -1,13 +1,12 @@
 package blackjack.factory;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Denomination;
+import blackjack.domain.card.Suit;
 import java.util.Arrays;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-
-import blackjack.domain.card.Card;
-import blackjack.domain.card.Denomination;
-import blackjack.domain.card.Suit;
 
 public class CardMockFactory {
 
@@ -16,7 +15,7 @@ public class CardMockFactory {
 	static {
 		cards = Arrays.stream(Denomination.values())
 				.map(denomination -> new Card(Suit.CLOVER, denomination))
-				.collect(Collectors.toMap(Card::getInformation, Function.identity()));
+				.collect(Collectors.toMap(Card::getDenominationAndSuit, Function.identity()));
 	}
 
 	public static Card of(final String information) {
